requirements.txt
anthropic>=0.45.0
streamlit>=1.42.0
qdrant-client>=1.13.0
sentence-transformers>=3.4.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
